[SafeMode] Last-minute fixes
